// Pawliner.Web/app/src/Models/OrderModel.ts
export default class OrderModel {
    id: number = 0;
    header: string = '';
    address: string = '';
    city: string = '';
    price: string = '';
    completedOn: string = '';
    description: string = '';
    createdAt: string = '';
    name: string = '';
    phoneNumber: string = '';
    status: number = 0;
    fileNmae: string = '';
    path: string = '';
    serviceClassiferId: string = '';
    serviceClassiferDescription: string = '';
    userId: string = '';
}